Agent_Frameworks/Pydantic_AI/requirements.txt
pydantic-ai[openai]>=1.0.0
openai>=1.0.0
httpx>=0.27.0
ddgs>=6.0.0
arxiv>=2.1.0
wikipedia>=1.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
rich>=13.0.0
